Clarify Evaluation Calculations (Addressing PR Comments)
Add pseudocode and commenting for the calculations involved in the
confusion matrix before accuracy, precision, and recall are evaluated.
Separate out the "none" case so that explicit false positives when there
are no gold boxes are singled out in a more readable format (just like
the way true negatives are). Refactor functions, and integrate
CorrectnessMetrics as a dataclass for readability