Also use upsampling for icons
Previously we were using all the points or just the keypoints from
icons, which works well in different cases. However, when there are both
small icons in a dataset, and large ones, using the same
upsampling/downsampling with icons is the most robust solution. We start
with the keypoints and upsample as needed.