Allow clients to pass in icon finder object
Clients can now pass in an already-instantiated icon finder object to
the benchmark pipeline instead of having to pass in a string option. The
benefit of this is that clients can instantiate the icon finder object
with whatever parameters they want (a dependency injection), and the
benchmark pipeline will simply call .find_icons.