Memory Calculation Updates
Memory_profiler's memory_usage invoked with a function passed into it
actually does not return the memory used by that function directly
(contrary to what their documentation implies). Rather, it's the memory
used by the entire current process in the duration of the function, in .1s intervals. Hence, to find the
memory used by one function, we configured the memory calculation to be
the maximum memory used during the function's execution subtracted by
the memory used right before executing the function.